refactor: evaluation/search 패키지 정리
